# docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/role": {
      "get": {
        "operationId": "RoleController_pagination",
        "parameters": [
          {
            "name": "order",
            "required": false,
            "in": "query",
            "schema": {
              "default": "DESC",
              "enum": [
                "ASC",
                "DESC"
              ],
              "type": "string"
            }
          },
          {
            "name": "orderBy",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "SYSTEM",
                "CUSTOM"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginationDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/RoleDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Roles"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "RoleController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Create role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Roles"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/role/permission": {
      "get": {
        "operationId": "RoleController_getAllPermissions",
        "parameters": [],
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Roles"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/role/my": {
      "get": {
        "operationId": "RoleController_myRoles",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Get role my",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RoleDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Roles"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/role/all": {
      "get": {
        "operationId": "RoleController_getAllUsers",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Get role list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RoleDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Roles"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/role/permissions": {
      "post": {
        "operationId": "RoleController_setPermissions",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleSetPermissionsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Create role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Roles"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/role/{roleId}": {
      "get": {
        "operationId": "RoleController_findOne",
        "parameters": [
          {
            "name": "roleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Roles"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "operationId": "RoleController_update",
        "parameters": [
          {
            "name": "roleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Update role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Roles"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "RoleController_delete",
        "parameters": [
          {
            "name": "roleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Delete role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "502": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionType"
                }
              }
            }
          }
        },
        "tags": [
          "Roles"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/role/permissions/{roleId}": {
      "get": {
        "operationId": "RoleController_getPermissions",
        "parameters": [
          {
            "name": "roleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get role",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RoleDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Roles"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/task": {
      "post": {
        "operationId": "TaskController_create",
        "parameters": [
          {
            "name": "project",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Create task",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Access denied"
          }
        },
        "tags": [
          "Tasks"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "TaskController_findAll",
        "parameters": [
          {
            "name": "boardId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "isBacklog",
            "required": true,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "project",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Create task",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Access denied"
          }
        },
        "tags": [
          "Tasks"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/task/{id}": {
      "get": {
        "operationId": "TaskController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "project",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Create task",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Access denied"
          }
        },
        "tags": [
          "Tasks"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "operationId": "TaskController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "project",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Create task",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Access denied"
          }
        },
        "tags": [
          "Tasks"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "TaskController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "project",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Delete task",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Access denied"
          }
        },
        "tags": [
          "Tasks"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Task Manager API",
    "description": "",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "PaginationDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "array"
            }
          },
          "totalCount": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          }
        },
        "required": [
          "data",
          "totalCount",
          "page",
          "limit"
        ]
      },
...
      "DeleteDto": {
        "type": "object",
        "properties": {
          "deleted": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "deleted"
        ]
      },
      "ExceptionType": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number"
          },
          "timestamp": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "validation": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "data": {
            "type": "object"
          }
        },
        "required": [
          "statusCode",
          "timestamp",
          "path",
          "message",
          "validation",
          "data"
        ]
      },
...
      "RoleDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "SYSTEM",
              "CUSTOM"
            ]
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "createdAt",
          "name",
          "type",
          "permissions"
        ]
      },
      "CreateRoleDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ]
      },
      "RoleSetPermissionsDto": {
        "type": "object",
        "properties": {
          "roleId": {
            "type": "string"
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "roleId",
          "permissions"
        ]
      },
      "UpdateRoleDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
...
      "IssueTypeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "icon": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "type": {
            "type": "string",
            "enum": [
              "Bug",
              "Task",
              "Sub-task",
              "Spike",
              "Test",
              "Research"
            ],
            "default": "Task"
          },
          "issueTypeColumns": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IssueTypeColumnDto"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "icon",
          "color",
          "isActive",
          "type",
          "issueTypeColumns",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      },
...
      "EpicDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "projectId": {
            "type": "number"
          },
          "position": {
            "type": "number"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "projectId",
          "position",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      },
      "TaskDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "issueTypeId": {
            "type": "number"
          },
          "issueType": {
            "$ref": "#/components/schemas/IssueTypeDto"
          },
          "epicId": {
            "type": "number"
          },
          "epic": {
            "$ref": "#/components/schemas/EpicDto"
          },
          "projectId": {
            "type": "number"
          },
          "project": {
            "$ref": "#/components/schemas/ProjectDto"
          },
          "boardId": {
            "type": "number"
          },
          "board": {
            "$ref": "#/components/schemas/BoardDto"
          },
          "boardColumnId": {
            "type": "number"
          },
          "boardColumn": {
            "$ref": "#/components/schemas/BoardColumnDto"
          },
          "isBacklog": {
            "type": "boolean"
          },
          "priority": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH"
            ],
            "default": "LOW"
          },
          "taskStatus": {
            "type": "string",
            "enum": [
              "ToDo",
              "InProgress",
              "Done"
            ],
            "default": "ToDo"
          },
          "assigneeId": {
            "type": "number"
          },
          "assignee": {
            "$ref": "#/components/schemas/UserDto"
          },
          "reporterId": {
            "type": "number"
          },
          "reporter": {
            "$ref": "#/components/schemas/UserDto"
          },
          "createdById": {
            "type": "number"
          },
          "createdBy": {
            "$ref": "#/components/schemas/UserDto"
          },
          "deletedById": {
            "type": "number"
          },
          "deletedBy": {
            "$ref": "#/components/schemas/UserDto"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "issueTypeId",
          "issueType",
          "epicId",
          "epic",
          "projectId",
          "project",
          "boardId",
          "board",
          "boardColumnId",
          "boardColumn",
          "isBacklog",
          "priority",
          "taskStatus",
          "assigneeId",
          "assignee",
          "reporterId",
          "reporter",
          "createdById",
          "createdBy",
          "deletedById",
          "deletedBy",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      },
...
      "CreateTaskPropertyDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "filedName": {
            "type": "string"
          },
          "value": {
            "type": "string"
          },
          "isRequired": {
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "filedName",
          "value",
          "isRequired"
        ]
      },
      "CreateTaskDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "issueTypeId": {
            "type": "number"
          },
          "taskProperty": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateTaskPropertyDto"
            }
          },
          "epicId": {
            "type": "number"
          },
          "boardId": {
            "type": "number"
          },
          "boardColumnId": {
            "type": "number"
          },
          "isBacklog": {
            "type": "boolean"
          },
          "priority": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH"
            ],
            "default": "LOW"
          },
          "taskStatus": {
            "type": "string",
            "enum": [
              "ToDo",
              "InProgress",
              "Done"
            ],
            "default": "ToDo"
          },
          "assigneeId": {
            "type": "number"
          },
          "reporterId": {
            "type": "number"
          }
        },
        "required": [
          "name",
          "description",
          "issueTypeId",
          "taskProperty",
          "boardId",
          "isBacklog",
          "priority",
          "taskStatus",
          "reporterId"
        ]
      },
...
      "UpdateTaskDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "issueTypeId": {
            "type": "number"
          },
          "taskProperty": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UpdateTaskPropertyDto"
            }
          },
          "epicId": {
            "type": "number"
          },
          "boardId": {
            "type": "number"
          },
          "boardColumnId": {
            "type": "number"
          },
          "isBacklog": {
            "type": "boolean"
          },
          "priority": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH"
            ],
            "default": "LOW"
          },
          "taskStatus": {
            "type": "string",
            "enum": [
              "ToDo",
              "InProgress",
              "Done"
            ],
            "default": "ToDo"
          },
          "assigneeId": {
            "type": "number"
          },
          "reporterId": {
            "type": "number"
          }
        }
      },
...
    }
  }
}